feat(UI): Updated the tailwindcss UI and pages directory
