--- frontend/src/components/css/navbar.css
@@ -1,83 +1,0 @@
-.navbar {
-  background-color: #247BA0;
-  padding: 10px 20px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1); 
-}
-
-.navbar-links {
-  display: flex;
-  align-items: center;
-  color: white;
-}
-
-.navbar-icon {
-  transform: translateY(2px);
-  margin-right: 2px;
-}
-
-.navbar-title {
-  font-size: 1.5rem;
-  text-decoration: none;
-  cursor: default;
-  color: white;
-}
-
-.navbar-link {
-  border-radius: 20px;
-  padding: 10px;
-  margin-left: 20px;
-  font-size: 1rem;
-  text-decoration: none;
-  color: white;
-  transition: all 0.3s ease; 
-  cursor: default; 
-  background-color: #70C1B3;
-}
-
-.navbar-link:hover {
-  background-color: #60cfbd;
-  transform: scale(1.02);
-  transition-property: background-color, transform;
-  transition: 0.4s ease-in-out;
-}
-
-.navbar-logout-icon {
-  height: 30px;
-  width: 30px;
-  color: white;
-}
-
-.navbar-logout {
-  position: relative; 
-  display: flex;
-  text-decoration: none;
-  cursor: default;
-  justify-content: flex-end;
-}
-
-.tooltip-logout {
-  visibility: hidden; 
-  width: 40px;
-  font-size: 12px;
-  background-color: white;
-  color: black;
-  text-align: center;
-  border-radius: 10px;
-  padding: 5px;
-  position: absolute;
-  z-index: 1;
-  top: calc(100% + 5px); 
-  left: 50%; 
-  transform: translateX(-50%);
-  opacity: 0;
-  transition: opacity 0.3s; 
-}
-
-.navbar-logout:hover .tooltip-logout {
-  visibility: visible; 
-  opacity: 1;
-}
-
