--- frontend/src/components/login/css/SignUp.css
@@ -1,94 +1,0 @@
-/* signupform.css */
-
-.SignUp {
-    display: flex;
-    flex-direction: column;
-    height: 600px;
-    width: 600px;
-    margin: 25px 0px 0px 500px;
-    padding: 10px;
-    border: 3px groove #ccc;
-    border-radius: 20px;
-  }
-  
-  .signupform-img{
-    margin: auto;
-    height: 50px;
-    width: 50px;
-  }
-  
-  .signupform-header{
-    font-size: 18px;
-    text-align: center;
-    margin: 10px 0px 25px 0px;
-  }
-  
-  .signupform-label{
-    width: 350px;
-    margin-left: 25px;
-    margin-bottom: 5px;
-    font-size: 16px;
-    font-weight: bold;
-  }
-  
-  .signupform-group {
-    display: inline-flex;
-    flex-direction: row;
-    margin: auto;
-    margin-bottom: 10px;
-    width: 500px;
-  }
-  
-  .signupform-control{
-    border: 2px inset lightskyblue;
-    font-size: 16px;
-    width: 500px;
-    height: 25px;
-    outline: none;
-    margin-left: 10px;
-    margin-right: 50px;
-    border-radius: 20px;
-    text-align: center;
-  }
-
-  .signupform-control-role{
-    font-size: 16px;
-    width: 400px;
-    height: 25px;
-    outline: none;
-    margin-left: 10px;
-    margin-right: 75px;
-    border-radius: 20px;
-    text-align: center;
-  }
-  
-  .signupform-control::placeholder{
-    text-align: center;
-  }
-  
-  .signupform-submit-container {
-    margin: auto;
-    position: relative;
-    display: inline-block; 
-  }
-  
-  .signupform-submit-btn {
-    position: relative;
-    z-index: 1; 
-    height: 50px;
-    width: 75px; 
-    text-align: center;
-    background-color: #70C1B3;
-    border-radius: 20px;
-    font-size: 16px;
-  }
-  
-  .signupform-submit-btn:hover{
-    background-color: #60cfbd;
-    transform: scale(1.02);
-    transition-property: background-color, transform;
-    transition: 0.4s ease-in-out;
-  }
-
-  
-
