--- frontend/src/components/login/css/login.css
@@ -1,144 +1,0 @@
-/* loginform.css */
-
-.LoginForm {
-  display: flex;
-  flex-direction: column;
-  height: 300px;
-  width: 500px;
-  transform: translate(100%,50%);
-  padding: 10px;
-  border: 3px groove #ccc;
-  border-radius: 20px;
-}
-
-.loginform-img{
-  margin: auto;
-  height: 50px;
-  width: 50px;
-}
-
-.loginform-header{
-  font-size: 18px;
-  text-align: center;
-  margin: 10px 0px 25px 0px;
-}
-
-.loginform-label{
-  width: 300px;
-  margin-left: 50px;
-  margin-bottom: 5px;
-}
-
-.loginform-group {
-  display: inline-flex;
-  flex-direction: row;
-  margin: auto;
-  width: 500px;
-}
-
-.loginform-control{
-  border: 2px inset lightskyblue;
-  font-size: 16px;
-  width: 500px;
-  height: 25px;
-  outline: none;
-  margin-right: 75px;
-  /* border: 3px inset blue; */
-  border-radius: 20px;
-  text-align: center;
-}
-
-.loginform-control::placeholder{
-  text-align: center;
-}
-
-.loginform-submit-container {
-  margin: auto;
-  position: relative;
-  display: inline-block; 
-}
-
-.loginform-submit-btn {
-  color: white;
-  position: relative;
-  z-index: 1; 
-  font-weight: 500;
-  height: 50px;
-  width: 75px; 
-  text-align: center;
-  background-color: #70C1B3;
-  border: none;
-  border-radius: 20px;
-  font-size: 16px;
-}
-
-.loginform-submit-btn:hover{
-  background-color: #60cfbd;
-  transform: scale(1.02);
-  transition-property: background-color, transform;
-  transition: 0.4s ease-in-out;
-}
-
-.loginform-error-msg{
-  margin-top: 5px;
-  text-align: center;
-  color: red;
-  font-weight: bold;
-  font-size: 16px;
-}
-
-.loginform-signup{
-  margin-top: 10px;
-  text-align: center;
-}
-
-.loader,
-.loader:before,
-.loader:after {
-  background: black;
-  animation: load1 1s infinite ease-in-out;
-  width: 1em;
-  height: 2em;
-}
-
-.loader {
-  width: 1em;
-  height: 2em;
-  color: black;
-  text-indent: -9999em;
-  margin: 0 auto;
-  margin-top: 3px;
-  position: relative;
-  font-size: 10px;
-  transform: translateZ(0);
-  animation-delay: -0.16s;
-}
-
-.loader:before,
-.loader:after {
-  position: absolute;
-  top: 0;
-  content: '';
-}
-
-.loader:before {
-  left: -1.5em;
-  animation-delay: -0.32s;
-}
-
-.loader:after {
-  left: 1.5em;
-}
-
-@keyframes load1 {
-  0%,
-  80%,
-  100% {
-    box-shadow: 0 0;
-    height: 3em;
-  }
-  40% {
-    box-shadow: 0 -2em;
-    height: 2em;
-  }
-}
