--- frontend/src/components/login/css/user.css
@@ -1,31 +1,0 @@
-/* user.css */
-
-.user-dropdown {
-    position: relative;
-  }
-  
-  .user-trigger {
-    display: inline-block;
-  }
-  
-  /* .user-content {
-    display: none;
-    position: absolute;
-    top: 100%;
-    left: 0;
-    background-color: #fff;
-    min-width: 200px;
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
-  } */
-  
-  .user-trigger:hover .user-content {
-    display: block;
-  }
-  
-  .user-content button {
-    display: block;
-    width: 100%;
-    padding: 10px;
-    text-align: center;
-  }
-
