+++ frontend/src/index.jsx
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
+import "tailwindcss/tailwind.css";
 import App from "./App";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
